nat: use proper type for counters
Type: improvement

+++ src/plugins/nat/nat.h
@@ -279,8 +279,8 @@ typedef struct
   u32 fib_index;
 /* *INDENT-OFF* */
 #define _(N, i, n, s) \
-  u16 busy_##n##_ports; \
-  u16 * busy_##n##_ports_per_thread; \
+  u32 busy_##n##_ports; \
+  u32 * busy_##n##_ports_per_thread; \
   u32 busy_##n##_port_refcounts[65535];
   foreach_nat_protocol
 #undef _
